Refactor: Mypage 클래스 분리
- LoadUserData
- MypageEventHandler
크게 두 클래스로 분리했는데 사실 분리된 클래스 내에서 메서드 역할이
여전히 좀 달라서 고민입니다.

GitHub issue #215
